Review content coherency and consistency - Module 14
Coherency audit findings and fixes:

HIGH IMPACT (fixed):
- H1: float16 → bfloat16 in solutions/03-multimodal-rag-solution.ipynb
  (CLIP model loading now consistent with other files)
- H2: float16 → bfloat16 in solutions/01-vision-language-demo-solution.ipynb
  (CLIP comparison function now consistent)

MEDIUM IMPACT (fixed):
- M1: Whisper input features dtype mismatch in 05-audio-transcription.ipynb
  (input features now use bfloat16 to match model dtype)
- M2: Same fix applied to scripts/audio_utils.py
- M3: Same fix applied to solutions/05-audio-transcription-solution.ipynb

All dtype now consistently bfloat16 for Blackwell optimization.
Removed obsolete fix_module_14.py script.
Updated COHERENCY_AUDIT_REPORT.md with complete findings.

+++ phase-3-advanced/module-14-multimodal/scripts/audio_utils.py
@@ -248,7 +248,7 @@ class AudioTranscriber:
             audio,
             sampling_rate=16000,
             return_tensors="pt"
-        ).input_features.to(self.model.device, dtype=torch.float16)
+        ).input_features.to(self.model.device, dtype=torch.bfloat16)
 
         forced_decoder_ids = None
         if language:
